Add fallback to gl
Replace gl alias with function gl(). The function allows us to access
the command line arguments ($@) and pass them along to the commands we
actually want to run.

First we try our custom Git alias
git lg $@

and if that does not exists we fall back to default Git behavior with

git log --oneline --graph $@

Note: This still has the disadvantage that if the alias "git lg" is not
defined, we get an error message in our output (before our actual
desired output). We should address this in a future iteration but for
now this is an improvement.

See #16