Replace gl function with alias
Replace the gl function with an alias pointing gl to
git lg

This maintains the current behavior if the Git alias
git lg
already exists.

However, if "git lg" does not exist, "gl" no longer
has a fallback of

git log --oneline --graph

and instead displays an error.

This change breaks backwards-compatibility (BC) and requires a
major version bump in the next release.

This change is documented in the CHANGELOG.md and on
https://github.com/salcode/salcode-zsh/wiki/Upgrade-Paths-for-Breaking-Releases#200

Resolves #42