Add alias pbpaste to ensure linebreak at end
When copying content from a textfield in a webpage and pasting it into a
file, the file is missing the ending linebreak (and this gets flagged by
Git).

The pbpasten alias calls pbpaste and then pipes the output through Perl
to ensure it ends with a linebreak.

pbpasten is "pbpaste" plus "n" (for new line)

Resolves #58